Seperated every screen and optimized code
