+++ lib/get_started_screen.dart
@@ -4,11 +4,13 @@ import 'package:personal_finance_tracker/login_page.dart';
 import 'package:personal_finance_tracker/signup_page.dart';
 
 class GetStartedScreen extends StatelessWidget {
+  const GetStartedScreen({super.key});
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Get Started'),
+        title: const Text('Get Started'),
       ),
       body: Center(
         child: Column(
@@ -19,33 +21,37 @@ class GetStartedScreen extends StatelessWidget {
               width: 200,
               height: 200,
             ),
-            SizedBox(height: 20),
-            Text(
+            const SizedBox(height: 20),
+            const Text(
               'Start your financial journey!',
               style: TextStyle(
                 fontSize: 24,
                 fontWeight: FontWeight.bold,
               ),
             ),
-            SizedBox(height: 20),
+            const SizedBox(height: 20),
             ElevatedButton(
               onPressed: () {
                 // Navigate to the signup screen
                 Navigator.of(context).pushReplacement(
-                  MaterialPageRoute(builder: (_) => SignUpPage()), // Replace `SignUpScreen` with the actual name of your signup screen
+                  MaterialPageRoute(
+                      builder: (_) =>
+                          const SignUpPage()), // Replace `SignUpScreen` with the actual name of your signup screen
                 );
               },
-              child: Text('Sign Up'),
+              child: const Text('Sign Up'),
             ),
-            SizedBox(height: 10),
+            const SizedBox(height: 10),
             TextButton(
               onPressed: () {
                 // Navigate to the login screen
                 Navigator.of(context).pushReplacement(
-                  MaterialPageRoute(builder: (_) => LoginPage()), // Replace `LoginPage` with the actual name of your login screen
+                  MaterialPageRoute(
+                      builder: (_) =>
+                          const LoginPage()), // Replace `LoginPage` with the actual name of your login screen
                 );
               },
-              child: Text('Log In'),
+              child: const Text('Log In'),
             ),
           ],
         ),
@@ -53,4 +59,3 @@ class GetStartedScreen extends StatelessWidget {
     );
   }
 }
-
