+++ lib/main.dart
@@ -5,24 +5,26 @@ import 'package:personal_finance_tracker/get_started_screen.dart';
 import 'package:personal_finance_tracker/login_page.dart';
 import 'package:personal_finance_tracker/home_screen.dart';
 
-
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
+      debugShowCheckedModeBanner: false,
       title: 'Personal Finance Tracker',
       initialRoute: '/', // Set the initial route
       routes: {
-        '/': (context) => SplashScreen(),
-        '/onboarding': (context) => OnboardingScreen(),
-        '/get_started': (context) => GetStartedScreen(),
-        '/login': (context) => LoginPage(),
-        'home': (context) => HomePage(),
-    // Add other screens here
+        '/': (context) => const SplashScreen(),
+        '/onboarding': (context) => const OnboardingScreen(),
+        '/get_started': (context) => const GetStartedScreen(),
+        '/login': (context) => const LoginPage(),
+        'home': (context) => const HomePage(),
+        // Add other screens here
       },
       theme: ThemeData(
         primarySwatch: Colors.blue,
@@ -30,4 +32,3 @@ class MyApp extends StatelessWidget {
     );
   }
 }
-
