+++ lib/signup_page.dart
@@ -4,11 +4,13 @@ import 'package:personal_finance_tracker/home_screen.dart';
 import 'package:personal_finance_tracker/login_page.dart';
 
 class SignUpPage extends StatelessWidget {
+  const SignUpPage({super.key});
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Sign Up'),
+        title: const Text('Sign Up'),
       ),
       body: Padding(
         padding: const EdgeInsets.all(16.0),
@@ -25,53 +27,53 @@ class SignUpPage extends StatelessWidget {
                 repeat: true, // Set to true if you want the animation to loop
               ),
               TextFormField(
-                decoration: InputDecoration(
+                decoration: const InputDecoration(
                   labelText: 'First Name',
                   icon: Icon(Icons.person),
                 ),
                 keyboardType: TextInputType.text,
               ),
-              SizedBox(height: 16.0),
+              const SizedBox(height: 16.0),
               TextFormField(
-                decoration: InputDecoration(
+                decoration: const InputDecoration(
                   labelText: 'Last Name',
                   icon: Icon(Icons.person),
                 ),
                 keyboardType: TextInputType.text,
               ),
-              SizedBox(height: 16.0),
+              const SizedBox(height: 16.0),
               TextFormField(
-                decoration: InputDecoration(
+                decoration: const InputDecoration(
                   labelText: 'Username',
                   icon: Icon(Icons.account_circle),
                 ),
                 keyboardType: TextInputType.text,
               ),
-              SizedBox(height: 16.0),
+              const SizedBox(height: 16.0),
               TextFormField(
-                decoration: InputDecoration(
+                decoration: const InputDecoration(
                   labelText: 'Phone Number',
                   icon: Icon(Icons.phone),
                 ),
                 keyboardType: TextInputType.phone,
               ),
-              SizedBox(height: 16.0),
+              const SizedBox(height: 16.0),
               TextFormField(
-                decoration: InputDecoration(
+                decoration: const InputDecoration(
                   labelText: 'Email',
                   icon: Icon(Icons.email),
                 ),
                 keyboardType: TextInputType.emailAddress,
               ),
-              SizedBox(height: 16.0),
+              const SizedBox(height: 16.0),
               TextFormField(
-                decoration: InputDecoration(
+                decoration: const InputDecoration(
                   labelText: 'Password',
                   icon: Icon(Icons.lock),
                 ),
                 obscureText: true,
               ),
-              SizedBox(height: 16.0),
+              const SizedBox(height: 16.0),
               ElevatedButton(
                 onPressed: () {
                   // Implement your signup logic here
@@ -79,27 +81,27 @@ class SignUpPage extends StatelessWidget {
                   // and handle the response accordingly.
                   // For this example, we'll show a success message.
                   ScaffoldMessenger.of(context).showSnackBar(
-                    SnackBar(
+                    const SnackBar(
                       content: Text('Sign Up Successful!'),
                     ),
                   );
 
                   // After successful signup, navigate to the login page
                   Navigator.of(context).pushReplacement(
-                    MaterialPageRoute(builder: (_) => HomePage()),
+                    MaterialPageRoute(builder: (_) => const HomePage()),
                   );
                 },
-                child: Text('Sign Up'),
+                child: const Text('Sign Up'),
               ),
-              SizedBox(height: 16.0),
+              const SizedBox(height: 16.0),
               TextButton(
                 onPressed: () {
                   // Navigate to the login page directly
                   Navigator.of(context).pushReplacement(
-                    MaterialPageRoute(builder: (_) => LoginPage()),
+                    MaterialPageRoute(builder: (_) => const LoginPage()),
                   );
                 },
-                child: Text('Already have an account? Log In'),
+                child: const Text('Already have an account? Log In'),
               ),
             ],
           ),
@@ -108,5 +110,3 @@ class SignUpPage extends StatelessWidget {
     );
   }
 }
-
-
